Move config maker into the main application, accessible now through command line arguments.
