+++ src/mapclient/core/exitcodes.py
@@ -1,5 +1,6 @@
 
 INVALID_WORKFLOW_LOCATION_GIVEN = 3
 HEADLESS_MODE_WITH_NO_WORKFLOW = 5
+CONFIGURATION_MODE_WITH_NO_DEFINITIONS = 6
 LOG_FILE_LOCK_FAILED = 7
 PID_FILE_LOCK_FAILED = 9
